iomete-airflow-plugin/pyproject.toml
[project]
name = "iomete-airflow-plugin"
version = "1.1.3"
description = "An Airflow plugin for interacting with IOMETE platform."
authors = [
    {name = "IOMETE"}
]
keywords = ["iomete", "airflow", "airflow plugin"]
packages = [{ include = "iomete_airflow_plugin" }]
license = "Apache Software License 2.0"
readme = "README.md"
requires-python = ">=3.9, <3.13"
dependencies = [
    "iomete-sdk (==2.1.2)",
    "Flask (==2.2.5)",
    "requests (==2.32.3)",
    "apache-airflow (==2.10.3)"

]

[tool.poetry.plugins."airflow.plugins"]
iomete = "iomete_airflow_plugin.plugin:IometePlugin"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.0.0"
bandit = "^1.7.5"
blue = ">=0.9.1, <1.0.0"
interrogate = ">=1.5.0, <2.0.0"
isort = ">=5.10.1,<7.0.0"
markdown = ">=3.4.0, <4.0.0"
mdformat-gfm = ">=0.3.5, <1.0.0"
mypy = "^1.8.0"
nbmake =  ">=1.4.3, <2.0.0"
setuptools = ">=80.0.0"
watchdog = "=6.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.interrogate]
color = true
exclude = ["setup.py", "docs", "build", "tests"]
fail-under = 10.0
generate-badge = "."
badge-format = "svg"
ignore-init-method = false
ignore-init-module = true
ignore-magic = true
ignore-module = true
ignore-nested-classes = true
ignore-nested-functions = true
ignore-private = true
ignore-property-decorators = true
ignore-regex = []
ignore-semiprivate = true
ignore-setters = true
omit-covered-files = true
quiet = false
verbose = 2
whitelist-regex = []

[tool.isort]
line_length = 120
profile = "black"
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]

[tool.blue]
line-length = 120